+++ src/operations/icc_transform.cc
@@ -47,7 +47,7 @@ VipsImage* PF::ICCTransformPar::build(std::vector<VipsImage*>& in, int first,
 				     VipsImage* imap, VipsImage* omap, 
 				     unsigned int& level)
 {
-  if( in.size() < first+1 ) {
+  if( (int)in.size() < first+1 ) {
     return NULL;
   }
