+++ src/operations/path_mask.cc
@@ -155,7 +155,7 @@ static void get_falloff_segments( const std::vector< std::pair<int,int> >& point
   int last0[2] = { -100, -100 }, last1[2] = { -100, -100 };
   //nbp = 0;
   int next = 0;
-  for(int i = 0; i < border.size(); i++) {
+  for(unsigned int i = 0; i < border.size(); i++) {
     p0[0] = points[i].first, p0[1] = points[i].second;
     if(next > 0)
       p1[0] = border[next].first, p1[1] = border[next].second;
@@ -163,7 +163,7 @@ static void get_falloff_segments( const std::vector< std::pair<int,int> >& point
       p1[0] = border[i].first, p1[1] = border[i].second;
 
     // now we check p1 value to know if we have to skip a part
-    if(next == i) next = 0;
+    if(next == (int)i) next = 0;
     while(p1[0] == -999999) {
       if(p1[1] == -999999)
         next = i - 1;
@@ -222,7 +222,7 @@ VipsImage* PF::PathMaskPar::build(std::vector<VipsImage*>& in, int first,
   falloff_curve.get().lock();
   if( falloff_curve.is_modified() || invert.is_modified() ) {
     //std::cout<<"PathMaskPar::build(): updating falloff LUT"<<std::endl;
-    for(int i = 0; i <= FormatInfo<unsigned short int>::RANGE; i++) {
+    for(unsigned int i = 0; i <= FormatInfo<unsigned short int>::RANGE; i++) {
       float x = ((float)i)/FormatInfo<unsigned short int>::RANGE;
       float y = falloff_curve.get().get_value( x );
       falloff_vec[i] = invert.get() ? 1.0f-y : y;
@@ -256,8 +256,8 @@ VipsImage* PF::PathMaskPar::build(std::vector<VipsImage*>& in, int first,
     const std::vector< std::pair<int,int> >& spline_points2 = get_smod().get_border();
 
     std::vector<path_point> path_points;
-    for( int pi = 0; pi < spline_points.size(); pi++ ) {
-      int pi2 = pi + 1;
+    for( unsigned int pi = 0; pi < spline_points.size(); pi++ ) {
+      unsigned int pi2 = pi + 1;
       if( pi2 >= spline_points.size() ) pi2 = 0;
       double px  = spline_points[pi].first,  py  = spline_points[pi].second;
       double px2 = spline_points[pi2].first, py2 = spline_points[pi2].second;
@@ -276,8 +276,8 @@ VipsImage* PF::PathMaskPar::build(std::vector<VipsImage*>& in, int first,
     //std::cout<<"PathMaskPar::build(): spline_points.size()="<<spline_points.size()<<"  path_points.size()="<<path_points.size()<<std::endl;
 
     int point_start = -1;
-    for( int pi = 0; pi < path_points.size(); pi++ ) {
-      int pi2 = pi + 1;
+    for( unsigned int pi = 0; pi < path_points.size(); pi++ ) {
+      unsigned int pi2 = pi + 1;
       if( pi2 >= path_points.size() ) pi2 = 0;
       int py = path_points[pi].y;
       int py2 = path_points[pi2].y;
@@ -293,7 +293,7 @@ VipsImage* PF::PathMaskPar::build(std::vector<VipsImage*>& in, int first,
         int pjprev = pj - 1;
         if( pjprev < 0 ) pjprev = path_points.size() - 1;
         int pjnext = pj + 1;
-        if( pjnext >= path_points.size() ) pjnext = 0;
+        if( pjnext >= (int)path_points.size() ) pjnext = 0;
 
         int px = path_points[pj].x;
         int py = path_points[pj].y;
@@ -310,7 +310,7 @@ VipsImage* PF::PathMaskPar::build(std::vector<VipsImage*>& in, int first,
         if( dy < 0 ) path_points[pj].state_changing = true;
 
         pj += 1;
-        if( pj >= path_points.size() ) pj = 0;
+        if( pj >= (int)path_points.size() ) pj = 0;
       } while( pj != point_start );
     }
 
@@ -324,7 +324,7 @@ VipsImage* PF::PathMaskPar::build(std::vector<VipsImage*>& in, int first,
     //std::cout<<"spline_points2.size(): "<<spline_points2.size()<<std::endl;
     int xlast=-1, ylast=-1;
     int yprev = -1000000;
-    for( int pi = 0; pi < path_points.size(); pi++ ) {
+    for( unsigned int pi = 0; pi < path_points.size(); pi++ ) {
       double px = path_points[pi].x, py = path_points[pi].y;
 
       int ipx = px, ipy = py;
@@ -333,7 +333,7 @@ VipsImage* PF::PathMaskPar::build(std::vector<VipsImage*>& in, int first,
       if( yprev == -1000000 ) yprev = ipy;
 
       int pi2 = pi + 1;
-      if( pi2 >= path_points.size() ) pi2 = 0;
+      if( pi2 >= (int)path_points.size() ) pi2 = 0;
 
       double px2 = path_points[pi2].x, py2 = path_points[pi2].y;
 
@@ -400,7 +400,7 @@ VipsImage* PF::PathMaskPar::build(std::vector<VipsImage*>& in, int first,
       //std::cout<<"PathMaskPar::build(): Filling segment "<<si<<"  ("<<y<<" -> "<<y+tw<<")"<<std::endl;
       //std::cout<<"  segments.size()="<<segments.size()<<std::endl;
       //if( segments.size()>0 ) std::cout<<"  segments[0]="<<segments[0].x1<<","<<segments[0].x1<<","<<segments[0].y1<<" -> "<<segments[0].x2<<","<<segments[0].y2<<std::endl;
-      for( int i = 0; i < segments.size(); i++ ) {
+      for( unsigned int i = 0; i < segments.size(); i++ ) {
         int ipx = segments[i].x1;
         int ipy = segments[i].y1;
         int ibpx = segments[i].x2;
